Add Display likes function
