+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "capstone-2",
+  "name": "CapstoneJS",
   "version": "1.0.0",
   "description": "Capstone Project",
   "main": "index.js",
