+++ src/index.js
@@ -1,10 +1,11 @@
 import './styles.css';
 import display from './modules/displayItems.js';
 import activateComments from './modules/comments.js';
-import likes from './modules/likes.js';
+import likes from './modules/countLikes.js';
 
 display()
   .then(() => {
     activateComments();
-    likes();
+    likes.displayLikes();
+    likes.activateLikeBtns();
   });
